src/index.ts
class PGSRenderer {
  private videoElement: HTMLVideoElement;
  private canvas: HTMLCanvasElement;
  private ctx: CanvasRenderingContext2D;
  private subtitles: { timestamp: number; imageData: ImageData }[] = [];
  private videoWidth: number;
  private videoHeight: number;

  constructor(videoElement: HTMLVideoElement, options: { skipDOMCheck?: boolean; videoWidth?: number; videoHeight?: number } = {}) {
    this.videoElement = videoElement;
    this.canvas = document.createElement('canvas');
    this.ctx = this.canvas.getContext('2d')!;
    this.videoWidth = options.videoWidth || this.videoElement.videoWidth;
    this.videoHeight = options.videoHeight || this.videoElement.videoHeight;
    
    if (!options.skipDOMCheck) {
      this.initCanvas();
    } else {
      this.initCanvasDimensions();
    }
  }

  private initCanvas() {
    if (!this.videoElement.parentElement) {
      throw new Error('Video element must be appended to the DOM before initializing PGSRenderer.');
    }
    this.initCanvasDimensions();
    this.videoElement.parentElement.appendChild(this.canvas);
  }

  private initCanvasDimensions() {
    this.canvas.width = this.videoWidth;
    this.canvas.height = this.videoHeight;
    this.canvas.style.position = 'absolute';
    this.canvas.style.top = '0';
    this.canvas.style.left = '0';
    this.canvas.style.pointerEvents = 'none'; // Make canvas transparent to pointer events
  }

  async loadSubtitles(subtitleData: ArrayBuffer) {
    this.subtitles = this.decodePGS(subtitleData);
    this.renderSubtitles();
  }

  private decodePGS(subtitleData: ArrayBuffer): { timestamp: number; imageData: ImageData }[] {
    // Placeholder decoding logic; replace with actual PGS decoding algorithm
    return [];
  }

  private renderSubtitles() {
    this.ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
    this.subtitles.forEach(subtitle => {
      const { timestamp, imageData } = subtitle;
      setTimeout(() => {
        this.ctx.putImageData(imageData, 0, 0);
      }, timestamp * 1000);
    });
  }
}

export { PGSRenderer };
